#69 adjusting ui, moving carousel
