--- angular-client/src/components/history/carousel.component.html
@@ -1,21 +1,0 @@
-<p-carousel [value]="runs" [numVisible]="1" [numScroll]="1" [circular]="true" [showIndicators]="false">
-  <ng-template let-run pTemplate="item">
-    <div class="runcard">
-      <div class="row">
-        <h3 class="left">Run #{{ run.id }}</h3>
-        <h3 class="right">Driver: {{ run.driverName }}</h3>
-      </div>
-      <div class="row">
-        <h3 class="left">{{ run.time }}</h3>
-        <h3 class="right">System: {{ run.systemName }}</h3>
-      </div>
-      <div class="row">
-        <h3 class="right">Location: {{ run.locationName }}</h3>
-      </div>
-    </div>
-  </ng-template>
-</p-carousel>
-<div class="vertical-center">
-  <button mat-button (click)="onNoClick()">Back</button>
-  <button mat-button (click)="onNoClick()" cdkFocusInitial>Select this Run</button>
-</div>
