+++ angular-client/src/components/history/history.component.html
@@ -1,1 +1,1 @@
-<button class="button" type="button" (click)="openDialog()">Historical</button>
+<button class="history-button" (click)="openDialog()">Historical</button>
