+++ angular-client/src/components/more-details/more-details.component.html
@@ -1,3 +1,1 @@
-<router-outlet></router-outlet>
-
 <button class="button" routerLink="/realtimedatagraph" type="button" ng-click="click(this)">More Details</button>
